% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Dataset}
\format{
## `fire`
A data frame with `n` rows and 2 columns:
\describe{
  \item{DISTANCE}{Numeric. Distance from a reference point (e.g., fire station) in unspecified units.}
  \item{DAMAGE}{Numeric. Amount of damage recorded, possibly in monetary terms.}
}
}
\source{
Internal dataset from `FIREDAM.csv`
}
\usage{
fire
}
\description{
A dataset containing information about fire damage and its relation to distance.
}
\keyword{datasets}
