% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinomial.R
\name{mybinomial}
\alias{mybinomial}
\title{Binomial Simulation Barplot}
\usage{
mybinomial(iter, n, p)
}
\arguments{
\item{iter}{An integer representing the number of iterations.}

\item{n}{An integer representing the number of trials per iteration.}

\item{p}{A numeric value between 0 and 1 representing the probability of success in each trial.}
}
\value{
A table of proportions representing the frequency of successes across all iterations.
}
\description{
Binomial Simulation Barplot
}
\examples{
\dontrun{mybin(iter = 100, n = 10, p = 0.5)}
}
