% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations. Default is 10000.}

\item{x}{A numeric vector of data values to be bootstrapped.}

\item{fun}{A function used to compute the statistic of interest.}

\item{alpha}{Significance level for the confidence interval. Default is 0.05.}

\item{cx}{Character expansion factor for text size in the plot. Default is 1.5.}

\item{...}{Additional graphical parameters to pass to the hist function}
}
\value{
An invisible list containing ci, fun, and x
}
\description{
This function performs bootstrap resampling to compute a confidence interval for a specified statistic.
}
\examples{
\dontrun{myboot2(iter = 10000, x = x, alpha = 0.05, fun = "mean", xlab = "Bootstrapped Means")}
}
