% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Generate Histogram of Sample Means Using CLT}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Integer. The sample size for each iteration.}

\item{iter}{Integer. The number of iterations (samples) to generate.}

\item{a}{Numeric. The lower bound of the uniform distribution (default is 0).}

\item{b}{Numeric. The upper bound of the uniform distribution (default is 10).}
}
\value{
A histogram of sample means with density curves overlaid.
}
\description{
Generate Histogram of Sample Means Using CLT
}
\examples{
\dontrun{mycltu(n=20,iter=100000)}
}
