% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve with Shaded Probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{A numeric value representing the mean of the normal distribution.}

\item{sigma}{A positive numeric value representing the standard deviation of the normal distribution.}

\item{a}{A numeric value representing the cutoff point for shading and probability calculation.}
}
\value{
A list containing the mean, standard deviation, cutoff value, and the probability.
}
\description{
Normal Curve with Shaded Probability
}
\examples{
\dontrun{myncurve(mu = 0, sigma = 1, a = 1.5)}
}
