% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson Maximum Likelihood Estimation}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{Starting value for the parameter.}

\item{delta}{Convergence threshold and step size for numerical derivative.}

\item{llik}{Log-likelihood function to be maximized.}

\item{xrange}{Range of x-values to use for plotting.}

\item{parameter}{Name of the parameter (used for labeling the x-axis).}
}
\value{
A list with vectors `x` (parameter estimates at each step) and `y` (derivatives at those steps).
}
\description{
Uses the Newton-Raphson method to find the value of a parameter that maximizes the log-likelihood function.
}
\examples{
\dontrun{myNRML(x0 = 1, delta = 0.0001,
       llik = function(lambda) log(dpois(4, lambda)),
       xrange = c(0, 10), parameter = "lambda")}
}
