% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Find Optimal Number of Tickets to Sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{An integer representing the total number of available seats.}

\item{gamma}{A numeric value representing the acceptable probability of overbooking.}

\item{p}{A numeric value representing the probability that a single ticketed passenger shows up.}
}
\value{
A list containing:
\item{nd}{Optimal number of tickets sold in the discrete case.}
\item{nc}{Optimal number of tickets sold in the continuous case.}
\item{N}{Total number of seats available.}
\item{p}{Probability of a passenger showing up.}
\item{gamma}{Acceptable probability of overbooking.}
}
\description{
Find Optimal Number of Tickets to Sell
}
\examples{
\dontrun{ntickets(N = 400, gamma = 0.02, p = 0.95)}
}
